requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
urllib3==1.25.10
xlrd==1.2.0
